+++ antlr/compile-all.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+# exit when any command fails
+set -e
+
+. compile-ss.sh
+. compile-plantuml.sh